Separated 'builtins' into 'Globals' and 'Syntax'.
'Syntax' replaces built-ins as the first object in the prototype chain
of every execution scope - so it contains flow control ("if" etc.),
important constants (eg. 'Nothing') and generally things that have to
be available quickly everywhere.

'Globals' is the parent of every module root, and contains everything
that is built-in, but should not shadow names from classes etc.