New bytecode flag: 'create property'.
This new flag is used to implement 'foo := 42;' inside a single opcode
instead of having to compile it as 'locals:foo = 42;" which requires
an additional fetch and function call. The flag could also be used for
special-casing the ':' operator, but this was not done at this point
to keep the special cases in the parser to a minimum.