Introduced argument sources and explicit lazy closure type to allow all the ways to make function calls to behave the same.
Originally, each way to make a call (LAZY opcode and vm_subcall_v) used their own logic to construct the execution scope. This is now a piece of common functionality.
Additionally, lazy closures are now a separate entity from interpreted functions, which allows checks like sepv_is_lazy(value) that are required by the new solution to argument passing.

This is a preparatory step - vm_subcall_v still has its own logic and is not switched to ArgumentSources yet.

until.sep had to be disabled since it's no longer possible to resolve() an arbitrary function (it only works on lazy values now) - it will be rewritten later when the invoke() method on function objects is added.