Flat function calls, e.g. 'return 42' now work - with some limitations.
'return a+b', 'return "Hi!"' and 'function argument' all work. However, it is impossible for the argument to start with an operator. So 'return [1]' doesn't work. The parser can't know that it should interpret this as 'return([1])' instead of 'return.[](1)' (indexing operator as in case of a[1]). In those cases, parentheses are necessary to disambiguate: 'return ([1,2,3])', 'return (-1)'.

Flat function calls also support only one argument - allowing anything else would result in code that is hard to read and parse for humans.